svm: move buildtimeCPUFeatures to CPUFeatureAccessImpl
