+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/aarch64/AArch64CPUFeatureAccess.java
@@ -27,7 +27,6 @@ package com.oracle.svm.core.aarch64;
 import java.util.ArrayList;
 import java.util.EnumSet;
 
-import org.graalvm.nativeimage.ImageSingletons;
 import org.graalvm.nativeimage.Platform;
 import org.graalvm.nativeimage.Platforms;
 import org.graalvm.nativeimage.StackValue;
@@ -35,7 +34,6 @@ import org.graalvm.nativeimage.c.struct.SizeOf;
 import org.graalvm.word.Pointer;
 
 import com.oracle.svm.core.CPUFeatureAccessImpl;
-import com.oracle.svm.core.SubstrateTargetDescription;
 import com.oracle.svm.core.UnmanagedMemoryUtil;
 import com.oracle.svm.core.annotate.Uninterruptible;
 import com.oracle.svm.core.util.VMError;
@@ -45,18 +43,9 @@ import jdk.vm.ci.code.Architecture;
 
 public class AArch64CPUFeatureAccess extends CPUFeatureAccessImpl {
 
-    private final EnumSet<?> buildtimeCPUFeatures;
-
     @Platforms(Platform.HOSTED_ONLY.class)
-    public AArch64CPUFeatureAccess() {
-        var targetDescription = ImageSingletons.lookup(SubstrateTargetDescription.class);
-        var arch = (AArch64) targetDescription.arch;
-        buildtimeCPUFeatures = EnumSet.copyOf(arch.getFeatures());
-    }
-
-    @Override
-    public EnumSet<?> buildtimeCPUFeatures() {
-        return buildtimeCPUFeatures;
+    public AArch64CPUFeatureAccess(EnumSet<?> buildtimeCPUFeatures) {
+        super(buildtimeCPUFeatures);
     }
 
     /**
